// src/types.d.ts
export interface Image {
  src: string;
  alt?: string;
}
// ...
export interface Project {
  /** Name of project */
  name: string;
  /** A short description of the project */
  description: string;
  /** An image used in the hero */
  img: Image;
  /** Link to project page or blog post */
  href: string;
}